64 bits fixes

+++ src/cmd/unit.cpp
@@ -627,7 +627,7 @@ void GameUnit<UnitType>::Draw(const Transformation &parent, const Matrix &parent
 	if (damagelevel<.99&&numm>0) {
 		unsigned int switcher=(damagelevel>.8)?1:
 			(damagelevel>.6)?2:(damagelevel>.4)?3:(damagelevel>.2)?4:5;
-		const unsigned int thus=(unsigned int)this;
+		const unsigned long thus=(unsigned long)this;
 		Mesh * tmp;
 		switch (switcher) {
 		case 5:
@@ -635,31 +635,31 @@ void GameUnit<UnitType>::Draw(const Transformation &parent, const Matrix &parent
 			if (tmp)
 				if (tmp->getBlendDst()==ONE)
 					tmp=meshdata[(thus+5+17)%(numm)];
-			LaunchOneParticle(*ctm,GetVelocity(),((int)this)+165,tmp,damagelevel,faction);
+			LaunchOneParticle(*ctm,GetVelocity(),((long)this)+165,tmp,damagelevel,faction);
 		case 4:
 			tmp=meshdata[(thus+47)%(numm)];
 			if (tmp)
 				if (tmp->getBlendDst()==ONE)
 					tmp=meshdata[(thus+47+17)%(numm)];			
-			LaunchOneParticle(*ctm,GetVelocity(),((int)this)+47,tmp,damagelevel,faction);
+			LaunchOneParticle(*ctm,GetVelocity(),((long)this)+47,tmp,damagelevel,faction);
 		case 3:
 			tmp=meshdata[(thus+61)%(numm)];
 			if (tmp)
 				if (tmp->getBlendDst()==ONE)
 					tmp=meshdata[(thus+47+61)%(numm)];						
-			LaunchOneParticle(*ctm,GetVelocity(),((int)this)+61,tmp,damagelevel,faction);
+			LaunchOneParticle(*ctm,GetVelocity(),((long)this)+61,tmp,damagelevel,faction);
 		case 2:
 			tmp=meshdata[(thus+2)%(numm)];
 			if (tmp)
 				if (tmp->getBlendDst()==ONE)
 					tmp=meshdata[(thus+1)%(numm)];						
-			LaunchOneParticle(*ctm,GetVelocity(),((int)this)+65537,tmp,damagelevel,faction);			
+			LaunchOneParticle(*ctm,GetVelocity(),((long)this)+65537,tmp,damagelevel,faction);			
 		default:
 			tmp=meshdata[(thus)%(numm)];
 			if (tmp)
 				if (tmp->getBlendDst()==ONE)
 					tmp=meshdata[(thus+129)%(numm)];
-			LaunchOneParticle(*ctm,GetVelocity(),((int)this)+257,tmp,damagelevel,faction);
+			LaunchOneParticle(*ctm,GetVelocity(),((long)this)+257,tmp,damagelevel,faction);
 		}
 	}
   }
